Rework handling of Montgomery constants
Instead m256_xxx() functions accepting the prime itself as a parameter,
make them take a structure containing the prime and associated
constants.

This saves a few bytes of code (52 with GCC, 20 with Clang) but most
importantly it avoids the dirty trick of relying on a particular limb of
the 2 primes used being 0 or 1, making the code more readable and more
generic (should anyone want to adapt it for another curve).