Rework use of CT_POISON()
For secret inputs that need to be decoded/validated, instead of
poisoning them only after decoding/validation, poison them from the
start and then unpoison the result of the validation.

This is a bit more secure as it prevents any accidental use before
poisoning, and it also resolves the ordering issues that were observed
with valgrind with the previous strategy.