Change internal mul64 abstraction
This paves the way for better assembly optimisation on some platforms.

For example, cores in the Cortex-M line with the DSP instruction set
extensions can implement this with a single instruction: umaal - however
the compiler doesn't seem to be taking advantage of it, so some assembly
seems called for.

Include tests, as assembly on cores that don't do it in a single
instruction will be non-trivial.