Add optimized u32_muladd64 for M0/M3/M23 cores
Benchmark total on a board based on an M0 core at 48 MHz: 9817 -> 5860.

That's -40% time per operation or +67% operations per unit of time.

That's also -72 bytes of code size and -48 bytes of stack usage.

Note: since we're manipulating the stack in asm inside a naked function,
the compiler has no way to know about the stack usage of that function,
so we need to fix the generated .su file manually with a hardcoded
value. Anyway, in a future commit I'll try using extended asm instead,
mostly to allow the compile to inline things (it can't with a naked
function), so this should go away at the same time.