Add optimized u32_muladd64 for M4/M7/M33 cores
Use extended asm rather than a naked function in order to allow
inlining - inspecting on the generated code shows GCC indeed inlines,
and nicely removes the redundant moves generated by the fact that the
function would need to put the return value in r1:r0.

Benchmark total on a board based on an M4 core at 100 MHz: 1525 -> 753.

That's -51% time per operation, or +102% operations per unit of time.

That's also -16 bytes of code, and few bytes less of stack usage.